update: print buttons of table pages for print all the data

+++ app/(plateform)/clients/bills/pay-dialog.tsx
@@ -13,7 +13,7 @@ import { Input } from "@/ui/input";
 type PayDialogProps = {};
 
 const schema = z.object({
-    amount: z.number().int().min(1),
+    amount: z.number().min(1),
 });
 
 export const PayDialog = ({}: PayDialogProps) => {
